requirements.txt
boto3
requests_futures
pandas
